Second Commit
Script modified to remove directory references to UCIHAR. Now assumed
that R script is in the folder where test and train are sub-folders.
Also, modified variable reference in the code - stdColumnNames and wrong
path for trainsubjects. Modified the CodeBook to remove the assumption
that I included before about UCIHAR directory.